Moved admin functions to own folder
After talking with emuLOAD about how the pages are currently set up, I
decided that instead of waiting until later to move all items that are
admin to a protected folder, I should do it now to minimize the issues
that could be caused later. So all pages remaining in "pages/" are
publically accessable in the end and all pages in "admin/" are
restricted access.