Cleanup on the parsenointro
Removed all of the labyrinine logic that tried to parse the HTML page
and replaced it with some preprocessing that strips the page of any text
or tags that we don't want to deal with. This leaves a very easy to
parse set that is then filtered by a simple array of known values. With
this, it can find all dumps on a page including bad and redumps.
Unbeknownst to me before, if a rom is bad, it has a tag that is called
"Section: Bad". Makes it easier than trying to parse the tag for "red".