Fix conditional inserts
Some games and disks don't have all of the information that is
"required" for a rom. This lead to an issue where a disk with a name and
sha1 only would not be imported because the information was considered
incomplete. With this new update, all files are handled conditionally:
whatever fields it has are added, and nothing else.