Fix DAT export
So, without even having to test this, I notice that in the db, that if
you don't provide a size (aka the file is empty OR it's a disk), then
the default value is 0, not "" or NULL. This means that if a file with
disks was exported, each disk would have a size "0" attribute added,
which is incorrect. It also means that files with no actual size
information will now not have size included. This might be useful for
some future imported sets.