Fixed generate again
I had the variables backward when it came to game naming. For some
reason, I had it that if the source was set, the source was written
alongside the name and if the manufacturer was set, THAT was written
alongside the name. Obviously, this is counterproductive. Found while
testing the FDS generate.