Clean up VideopacNL
Finally took time to understand how this checker worked. Turns out that
literally half of the code was dedicated to using a proxy to access the
site. The proxy does not seem necessary (and flat out errors out when I
tried to use it) so all of that code has been removed. On the other
side, I figured out which cookie was used, how the cookie was being used
by the function and how it was being checked. WoD had hardcoded a value
to check for which has now been replaced by a preg_match.