Header notes and todo update
Added a few notes on the header files that I was able to run across,
including which ones would be absolutely necessary and which ones are
duplicates. Duplicates or functional subsets are commented out but
preserved. Also updated the TODOs to include instructions on how to
implement auto-gen as well as a note about the headers.