Create packaging for create all DATs
If we're in auto mode, we assume that it's for the purposes of sharing,
so a folder full of unsorted DATs is pretty useless. This code creates a
zipfile full of the newest DATs, sorted into 3 folders: custom (if
system and source are set), merged-system (for when the system is set
only), and merged-source (for when the source is set only). It currently
doesn't auto-create or archive the MEGAMERGED because of memory
problems.