Update database backup
I didn't mention it in previous backups, but these do NOT have duplicate
data removed quite yet. Once I have everything in, THEN the database
will be cleaned. This is to make sure that nothing is being removed that
shouldn't be (even though the script worked fine on a smaller set of
data).