Fix XML-derived DAT creation
Users mentioned that some DATs that contain special characters were
throwing errors when being used in RV and CMP. It turns out that the
game and rom names needed to be html encoded so that the program would
recognize them properly. Waiting for feedback to see if any other fixes
are necessary.