Enable individual DATs to be downloaded
Based on the work by emuLOAD, if a DAT isn't being created en-masse, it
will be automatically gzipped and given to the user as a download. Since
this is a nested page, some changes had to be made to the parent page as
well, but nothing that should impact other pages.