Limit DAT creation for certain sources
For all import-only sources, there should be no custom DAT created as
this is stepping directly on the toes of the people who went through the
work of making them in the first place. Coincidentally, sources with
values of 14 and lower are all import-only, and thus, makes it easy to
filter by. If more import-only DATs are added in the future, an explicit
array should be used.