Add N64 header file
Even though it's not used by anything currently, having this "ABCD"
header is a really good idea, especially for the merged N64 DAT. This
was mostly a concern because of the new GoodN64 set that just got
released. The best course of action is to convert all to V64, hash it,
and then check against byte- and word-swapped roms to see that they
still match.