Add Deheadering page
This new page aids in creating a deheadered set by checking an entire
folder for a set of headers and then making a copy of the file without a
header if one is found. This can be extended for any header that will
have a <header>.xml file associated with it. Currently, all header skips
are included and have been tested on local sets.