Experimental No-Intro parsing update
I realized that the code that was being used here was suspiciously
similar to how the code in the online checkers was being handled. As
such, I was using things like explode and implode and going line by line
to get data. Once I was able to track my own code again, it was updated
to using regex, which creates an actual usable output (currently only
accepted and scene, no redumps).