Add further functionality to experimental No-Intro parsing
Now that the parsing itself works properly, I added a simple DAT output
based on a slightly hacked out version of the generate code. As this
test page progresses, it may end up being able to be added to "sites/"
as the online checker for No-Intro. Next steps include adding redump
addition, adding a check for system id (instead of the hardcoded NDS id
used currently), adding a generic dropdown page for selecting a system,
and removing (or commenting out) the auto-generate mapping code.