Clean up computeremuzone online checker
This one is a bit special. First and foremost, it had some of the most
explode cleanup that I've ever run into. Second, I found a nifty little
function called pathinfo which can get the extension from a filename
without using explode. This will come in handy as time goes on. (This is
why AmstradESP was updated too.) Lastly, because of how the previous
version was programmed, the associated txt had to be updated to scrub
some of the names (which had parts of the link appended by accident).