Simplify process_roms code
The code that was being used before was horrendously complex. First it
got the roms as an array sorted by machine and file name. Then it
renamed the machines and files as necessary. Then it sorted the entire
set if it was merged and got rid of duplicates (not even properly! It
disregarded the size completely!). Finally, it REsorted the entire array
to be in machine / file name order.

The code is now much simpler. It first retrieves the data with a
stipulation: if we're in a merged mode, it automatically gets rid of any
duplicates based on size, crc, md5, sha1. Then it does a much easier
sort to machine / file name order. Finally, it traverses the output and
renames anything appropriate. The code is about half the size and a
billion times more effective (hyperbole).

Next up is updating the other part of the generate code, if possible.