Ensure that all hashes are uppercase
It turns out that RV is VERY particular about the hashes that it is able
to parse. I've been lucky in the testing so far that this has not come
up, but this is actually a backport from the desktop version where
nombre002/@tractivo found this issue and reported it.