Possibly fix erroneous merging
This is another backport from the desktop version which is somewhat
based on feedback from nombre002/@tractivo. In another turn of
coincidence so far, all WoD custom DATs (the testing ground for most of
the app) all have CRC, MD5, and SHA1. This is not the case for all sets;
some have only CRC and MD5 or CRC and SHA1 (or even just CRC). With the
original code, these duplicates would be left in because they didn't
match in every single aspect. This is a tenative fix that may have
further fixes that are coordinated with the desktop version in the
future.