Rewrite merging code
The code that is back here is loosely based on the original 3 step code
from previous iterations. This is a backport of the optimized code used
in the desktop version. Here, there is only at most one sort as the
query gets the correct sorting order for merged and unmerged. This code
needed to be implemented because of the varying degrees of available
data on any given ROM from different sources. Once other sources (such
as TOSEC) begin to be added, this will become apparent.