Backport automatic type detection from desktop version
The desktop version had smart code that allowed it to figure out
(usually) the type of the DAT based on its filename pattern. The
previous PHP code used a param to deterime the type of the DAT, which,
though it's more distinctive, leads to having different folders for each
type, and doesn't allow all types to be imported at once. The new code
here uses the smart sorting to determine filetype and then uses that
type as if it were a param.