Overhaul parsenointro
ParseNoIntro is now a very specialized page. Instead of trying to get
ALL information from a page, it just captures the scene release info
instead. To make it even more efficient, it uses an XML reader to strip
out a lot of the junk, along with a couple levels of stripping
unnecessary whitespace.

The result is a very robust page that can traverse and get information
for Nintendo DS ONLY. Further tests will be done for 3DS in the future.
For now, it will remain limited until this information can be written to
file, which I hope will come in an update soon.