Fix rv2xml
It turns out that I needed to understand why my regex was failing in
order to figure out why the XML wasn't writing out. The XML writer was
being initialized improperly, wasn't having some settings set, some of
the regex was wrong, lines needed to be trimmed, and logic needed to be
backported for rom/disk handling.